created make_dipole_forward in _make_forward.py
fixed missing import
