+++ mne/simulation/source.py
@@ -7,12 +7,10 @@
 import numpy as np
 
 from ..source_estimate import SourceEstimate, VolSourceEstimate
-from ..source_space import (_ensure_src, _make_discrete_source_space,
-                            SourceSpaces)
+from ..source_space import _ensure_src
 from ..utils import check_random_state, warn
 from ..externals.six.moves import zip
-from ..forward import make_forward_solution, convert_forward_solution
-from ..fixes import in1d
+
 
 def select_source_in_label(src, label, random_state=None):
     """Select source positions using a label
@@ -220,121 +218,3 @@ def simulate_stc(src, labels, stc_data, tmin, tstep, value_fun=None):
 
     stc = SourceEstimate(data, vertices=vertno, tmin=tmin, tstep=tstep)
     return stc
-
-
-def make_forward_dipole(dipole, bem, info, trans=None, n_jobs=1, verbose=None):
-    """Convert dipole object to source estimate and calculate forward operator
-
-    The instance of Dipole is converted to a discrete volume source space,
-    which is then combined with a BEM (can also be a sphere model dict) and
-    the sensor information in info to form a forward operator.
-
-    The source estimate object (with the forward operator) can be projected to
-    sensor-space using :func:`mne.simulation.evoked.simulate_evoked`.
-
-    Parameters
-    ----------
-    dipole : Dipole object
-        Instance of Dipole containing position, orientation and amplitude of
-        one or more dipoles. Several dipoles may occur simultaneously, but
-        when more than one time point is defined, the spacing between them
-        (i.e. the sampling rate) must be constant.
-    bem : str | dict
-        The BEM filename (str) or a loaded sphere model (dict).
-    info : dict
-        Sensor-info etc., e.g., from real evoked file.
-    trans : str | None
-        The head<->MRI transform filename. Must be provided unless BEM
-        is a sphere model.
-    n_jobs : int
-        Number of jobs to run in parallel (used in making forward solution).
-    verbose : bool, str, int, or None
-        If not None, override default verbose level (see mne.verbose).
-
-    Returns
-    -------
-    stc : instance of VolSourceEstimate
-        The dipoles converted to a discrete set of points and associated
-        time courses.
-    fwd : instance of Forward
-        The forward solution.
-
-    See Also
-    --------
-    mne.simulation.simulate_evoked
-
-    Notes
-    -----
-    .. versionadded:: 0.12.0
-    """
-    # Make copies to avoid mangling original dipole
-    times = dipole.times.copy()
-    pos = dipole.pos.copy()
-    amplitude = dipole.amplitude.copy()
-    ori = dipole.ori.copy()
-
-    # Convert positions to discrete source space (allows duplicate rr & nn)
-    # NB information about dipole orientation enters here, then no more
-    sources = dict(rr=pos, nn=ori)
-    # Dipole objects must be in the head frame
-    sp = _make_discrete_source_space(sources, coord_frame='head')
-    src = SourceSpaces([sp])  # dict with working_dir, command_line not nec
-
-    # Forward operator created for channels in info (use pick_info to restrict)
-    # Use defaults for most params, including min_dist
-    fwd = make_forward_solution(info, trans, src, bem, fname=None,
-                                n_jobs=n_jobs, verbose=verbose)
-    # Convert from free orientations to fixed (in-place)
-    convert_forward_solution(fwd, surf_ori=False, force_fixed=True,
-                             copy=False, verbose=None)
-
-    # Check for omissions due to proximity to inner skull in
-    # make_forward_solution, which will result in an exception
-    if fwd['src'][0]['nuse'] != len(pos):
-        inuse = fwd['src'][0]['inuse'].astype(np.bool)
-        head = ('The following dipoles are outside the inner skull boundary')
-        msg = len(head)*'#' + '\n' + head + '\n'
-        for (t, pos) in zip(times[np.logical_not(inuse)],
-                            pos[np.logical_not(inuse)]):
-            msg += '    t={:.0f} ms, pos=({:.0f}, {:.0f}, {:.0f}) mm\n'.\
-                format(t*1000., pos[0]*1000., pos[1]*1000., pos[2]*1000.)
-        msg += len(head)*'#'
-        print(msg)
-        raise ValueError('One or more dipoles outside the inner skull')
-
-    # multiple dipoles (rr and nn) per time instant allowed
-    # uneven sampling in time returns list
-    timepoints = np.unique(times)
-    if len(timepoints) > 1:
-        tdiff = np.diff(timepoints)
-        if not np.allclose(tdiff, tdiff[0]):
-            warn('Unique time points of dipoles unevenly spaced: returned '
-                 'stc will be a list, one for each time point.')
-            tstep = -1.0
-        else:
-            tstep = tdiff[0]
-    elif len(timepoints) == 1:
-        tstep = 0.001
-
-    # Build the data matrix, essentially a block-diagonal with
-    # n_rows: number of dipoles in total (dipole.amplitudes)
-    # n_cols: number of unique time points in dipole.times
-    # amplitude with identical value of times go together in one col (others=0)
-    data = np.zeros((len(amplitude), len(timepoints)))  # (n_d, n_t)
-    row = 0
-    for tpind, tp in enumerate(timepoints):
-        amp = amplitude[in1d(times, tp)]
-        data[row:row+len(amp), tpind] = amp
-        row += len(amp)
-
-    if tstep > 0:
-        stc = VolSourceEstimate(data, vertices=fwd['src'][0]['vertno'],
-                                tmin=timepoints[0],
-                                tstep=tstep, subject=None)
-    else:  # Must return a list of stc, one for each time point
-        stc = []
-        for col, tp in enumerate(timepoints):
-            stc += [VolSourceEstimate(data[:, col][:, np.newaxis],
-                                      vertices=fwd['src'][0]['vertno'],
-                                      tmin=tp, tstep=0.001, subject=None)]
-    return stc, fwd
